+++ src/main/java/com/silashkin/telegram/cache/UserCache.java
@@ -7,15 +7,15 @@ import org.springframework.stereotype.Component;
 @Component
 public class UserCache implements Cache {
 
-    private final HashMap<Integer, HandlerInterface> userHandlers = new HashMap<>();
+    private final HashMap<Long, HandlerInterface> userHandlers = new HashMap<>();
 
     @Override
-    public HandlerInterface getCacheHandler(int i) {
+    public HandlerInterface getCacheHandler(long i) {
         return userHandlers.get(i);
     }
 
     @Override
-    public void setCacheHandler(int i, HandlerInterface handler) {
+    public void setCacheHandler(long i, HandlerInterface handler) {
         userHandlers.put(i, handler);
     }
 }
